# frontend/src/lib/openapi.yaml
openapi: 3.0.0
info:
  description: API for managing blips, technologies, users, and user technologies
  title: Tech Radar API
  contact: {}
  version: "1.0"
paths:
  /auth/login:
    post:
      description: Authenticate user and return JWT token
      tags:
        - auth
      summary: Login user
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/dto.LoginRequest"
        description: Login credentials
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/dto.AuthResponse"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/handlers.Error"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/handlers.Error"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/handlers.Error"
...
  /auth/register:
    post:
      description: Create a new user account and return authentication token
      tags:
        - auth
      summary: Register a new user
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/dto.RegisterRequest"
        description: Registration data
        required: true
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/dto.AuthResponse"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/handlers.Error"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/handlers.Error"
  /blips:
    get:
      description: Get all blips
      tags:
        - blips
      summary: Get all blips
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/handlers.Blip"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/handlers.Error"
    post:
      description: Create a new blip
      tags:
        - blips
      summary: Create a blip
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/handlers.CreateBlipRequest"
        description: Blip data
        required: true
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/handlers.Blip"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/handlers.Error"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/handlers.Error"
  "/blips/{id}":
    get:
      description: Get blip by ID
      tags:
        - blips
      summary: Get a blip
      parameters:
        - description: Blip ID
          name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/handlers.Blip"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/handlers.Error"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/handlers.Error"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/handlers.Error"
    put:
      description: Update blip by ID
      tags:
        - blips
      summary: Update a blip
      parameters:
        - description: Blip ID
          name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/handlers.UpdateBlipRequest"
        description: Blip data
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/handlers.Blip"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/handlers.Error"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/handlers.Error"
    delete:
      description: Delete blip by ID
      tags:
        - blips
      summary: Delete a blip
      parameters:
        - description: Blip ID
          name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/handlers.Error"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/handlers.Error"
...
servers:
  - url: //localhost:8080
components:
  schemas:
    dto.AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/dto.UserDTO"
    dto.LoginRequest:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
    dto.RegisterRequest:
      type: object
      properties:
        email:
          type: string
        name:
          type: string
        password:
          type: string
        username:
          type: string
    dto.UserDTO:
      type: object
      properties:
        email:
          type: string
        id:
          type: string
        is_admin:
          type: boolean
        name:
          type: string
        username:
          type: string
    handlers.Blip:
      type: object
      properties:
        context: {}
        created_at:
          type: string
        id:
          type: integer
        updated_at:
          type: string
    handlers.CreateBlipRequest:
      type: object
      properties:
        context:
          type: object
          additionalProperties: true
...
    handlers.Error:
      type: object
      properties:
        message:
          type: string
...
    handlers.UpdateBlipRequest:
      type: object
      properties:
        context:
          type: object
          additionalProperties: true
...